Splitout enums and obsolute signatures.py
